Predict deployable barriers
